fix hand axe in cube recipes
in some recipes hand axe was rolled as output was specified as item type without lvl 99 and no proper item of that type and level was found.
Added missing lvl 99 to some recipes. maybe even to those that didnt really need it.
Removed some cube recipes for uniques that do no longer exist